Initial version
After cleaning up the large `atom` folder.

Atom plugins can be installed like this:

    apm install --packages-file atom-package-list.txt

**Please checkout a fresh copy of this repository because the history
has been rewritten.**

*DEVELOPER NOTE*
The large `atom` folder has been removed because the repository got out
of bound and was hard to maintain and use. Especially on slow networks.

So the repository is having a fresh start! I also removed settings for
tools I didn't use in the recent year(s).

Best is to `fork` the repository and make it your own.

Sorry for any inconvenience caused.