README: update content to reflect the latest workflow
The main goal is to have a simple and quick way to setup a new
workstation.

Turned out that I seldomly use Ubuntu as a workstation and macOS is my
most common working environment.

Hence I drop the support for Ubuntu to keep the file easy to gasp and
understand i.e. simply copy a few commands and I'm done.